chore: separate env UI (#489)
* feat: refactor UI components to separate packages
