+++ common/changes/@aws/workbench-core-accounts-ui/chore-separate-env-UI_2022-09-14-23-16.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "packageName": "@aws/workbench-core-accounts-ui",
+      "comment": "feat: separate out UI packages",
+      "type": "minor"
+    }
+  ],
+  "packageName": "@aws/workbench-core-accounts-ui"
+}
