+++ common/config/rush/browser-approved-packages.json
@@ -127,6 +127,10 @@
       "allowedCategories": [ "production" ]
     },
     {
+      "name": "@aws/workbench-core-environments-ui",
+      "allowedCategories": [ "production" ]
+    },
+    {
       "name": "@aws/workbench-core-example-express",
       "allowedCategories": [ "prototypes" ]
     },
@@ -139,7 +143,15 @@
       "allowedCategories": [ "production", "prototypes" ]
     },
     {
+      "name": "@aws/workbench-core-accounts-ui",
+      "allowedCategories": [ "production" ]
+    },
+    {
       "name": "@aws/workbench-core-swb-common-ui",
+      "allowedCategories": [ "production" ]
+    },
+    {
+      "name": "@aws/workbench-core-swb-datasets-ui",
       "allowedCategories": [ "production" ]
     },
     {
