+++ rush.json
@@ -108,6 +108,18 @@
       "shouldPublish": true
     },
     {
+      "packageName": "@aws/workbench-core-environments-ui",
+      "projectFolder": "workbench-core/environments-ui",
+      "reviewCategory": "production",
+      "shouldPublish": true
+    },
+    {
+      "packageName": "@aws/workbench-core-accounts-ui",
+      "projectFolder": "workbench-core/accounts-ui",
+      "reviewCategory": "production",
+      "shouldPublish": true
+    },
+    {
       "packageName": "@aws/workbench-core-example-express",
       "projectFolder": "workbench-core/example/express",
       "reviewCategory": "prototypes",
