+++ workbench-core/accounts-ui/.eslintignore
@@ -1,0 +1,4 @@
+.eslintrc.js
+node_modules
+lib
+temp
