+++ workbench-core/accounts-ui/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@aws/swb-ui",
+  "name": "@aws/workbench-core-accounts-ui",
   "version": "0.0.1",
-  "description": "NextJS frontend UI for Service Workbench v2",
+  "description": "Package that contains UI components for Projects",
   "homepage": "https://github.com/awslabs/solution-spark-on-aws",
   "repository": {
     "type": "git",
@@ -12,77 +12,53 @@
     "name": "Amazon Web Services",
     "url": "http://aws.amazon.com"
   },
-  "main": "lib/index.js",
-  "typings": "lib/index.d.ts",
+  "main": "lib/esm/index.js",
+  "module": "./lib/esm/index.js",
+  "typings": "lib/esm/index.d.ts",
   "scripts": {
     "add-license-header": "license-check-and-add add -f license-add-config.json",
-    "build": "next build && rushx lint",
-    "build:test": "rushx build && rushx test",
+    "build": "heft build --clean && rushx pkg-json-lint",
+    "build:test": "heft test --clean && rushx pkg-json-lint && rushx make-badges",
     "check-license-header": "license-check-and-add check -f license-add-config.json",
     "depcheck": "depcheck",
-    "deploy-ui-and-api": "bash ./scripts/deploy.sh",
-    "dev": "bash ./scripts/dev.sh",
-    "export": "next build && next export",
     "license-checker": "license-checker --onlyAllow 'MIT; Apache-2.0; ISC; BSD'",
     "lint": "eslint . && rushx pkg-json-lint",
     "lint:fix": "eslint . --fix",
-    "make-badges": "istanbul-badges-readme --coverageDir=./temp/coverage",
+    "make-badges": "istanbul-badges-readme --coverageDir=./temp/coverage --exitCode=1",
     "pkg-json-lint": "npmPkgJsonLint -c ../../.npmpackagejsonlintrc.json .",
-    "serve": "node .next/server/webpack-runtime.js",
     "sort-package-json": "sort-package-json package.json",
-    "test": "jest --coverage --passWithNoTests 2>&1 && rushx make-badges",
-    "test:only": "jest --ci --coverage --passWithNoTests 2>&1"
+    "test": "rushx test:only && rushx make-badges",
+    "test:only": "heft test --clean --no-build"
   },
   "dependencies": {
-    "@aws/workbench-core-datasets-ui": "workspace:*",
-    "@aws/workbench-core-environments-ui": "workspace:*",
-    "@aws/workbench-core-environments": "workspace:*",
     "@aws/workbench-core-swb-common-ui": "workspace:*",
     "next": "^12.1.5",
-    "next-global-css": "^1.2.0",
-    "next-transpile-modules": "^9.0.0",
     "react": "17.0.2",
     "react-dom": "17.0.2",
-    "swr": "^1.3.0"
+    "swr": "^1.3.0",
+    "typescript": "^4.5.2"
   },
   "devDependencies": {
     "@aws/eslint-config-workbench-core-eslint-custom": "workspace:*",
-    "@babel/core": "^7.17.9",
-    "@babel/preset-env": "^7.16.11",
-    "@babel/preset-react": "^7.16.7",
-    "@babel/preset-typescript": "^7.16.7",
     "@cloudscape-design/collection-hooks": "^1.0.3",
     "@cloudscape-design/components": "^3.0.27",
-    "@cloudscape-design/design-tokens": "^3.0.1",
-    "@cloudscape-design/global-styles": "^1.0.1",
     "@rushstack/eslint-config": "^3.0.0",
     "@rushstack/heft": "^0.47.5",
     "@rushstack/heft-jest-plugin": "^0.3.28",
     "@rushstack/heft-node-rig": "^1.10.5",
-    "@testing-library/dom": "^8.13.0",
-    "@testing-library/jest-dom": "5.16.1",
-    "@testing-library/react": "12.1.2",
-    "@testing-library/user-event": "13.5.0",
     "@types/heft-jest": "1.0.2",
-    "@types/jest": "^27.4.1",
     "@types/node": "^14",
     "@types/react": "17.0.38",
-    "babel-jest": "27.5.1",
+    "@types/react-dom": "^18.0.6",
     "depcheck": "^1.4.3",
     "eslint": "^8.7.0",
-    "eslint-config-next": "12.1.5",
     "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-testing-library": "5.0.1",
     "istanbul-badges-readme": "1.8.1",
-    "jest": "^27.5.1",
     "license-check-and-add": "^4.0.5",
     "license-checker": "^25.0.1",
     "npm-package-json-lint": "^6.3.0",
     "npm-package-json-lint-config-default": "^5.0.0",
     "sass": "^1.49.8",
-    "sharp": "^0.30.4",
-    "sort-package-json": "^1.57.0",
-    "ts-jest": "^27.1.3",
-    "typescript": "^4.5.2"
+    "sort-package-json": "^1.57.0"
   }
 }
