+++ workbench-core/environments-ui/src/api/environmentTypeConfigs.ts
