+++ workbench-core/environments-ui/src/environments-table-config/workspacesFilteringOptions.tsx
