+++ workbench-core/environments-ui/src/index.tsx
@@ -1,0 +1,12 @@
+/*
+ *  Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *  SPDX-License-Identifier: Apache-2.0
+ */
+
+import EnvironmentsPage from './pages/environments/index';
+import NewEnvironmentPage from './pages/environments/new';
+
+
+export {
+  EnvironmentsPage, NewEnvironmentPage
+};
